Added mention of building forntend
Listen, I know that it's pull request and I shouldn't write something wierd here, but I will. I think I have a rights to do that just because I was trying to do fix the issue for 4 hours and the issue was in me. I just don't want anyone else to make the same mistake and spend 4 or even more hours on that